+++ tests/key-sequence.js
@@ -212,6 +212,76 @@ $(document).ready(function() {
 
 	});
 
+	module("quirk-detection", {
+		
+	});
+	
+	asyncTest("quirk detection", 1, function() {
+		setTimeout(function() {
+			var iFrameDoc = window.frames[0].document;
+	
+			var jquery = iFrameDoc.createElement('script');
+			jquery.src = "../libs/jquery-1.10.2.js";
+			iFrameDoc.body.appendChild(jquery);
+	
+			var bililiteRange = iFrameDoc.createElement('script');
+			bililiteRange.src = "../libs/bililiteRange.js";
+			iFrameDoc.body.appendChild(bililiteRange);
+			
+			var jquerySimulate = iFrameDoc.createElement('script');
+			jquerySimulate.src = "../libs/jquery.simulate.js";
+			iFrameDoc.body.appendChild(jquerySimulate);
+			
+			var simulateExt = iFrameDoc.createElement('script');
+			simulateExt.src = "../src/jquery.simulate.ext.js";
+			iFrameDoc.body.appendChild(simulateExt);
+	
+			var simulateExtKeySequence = iFrameDoc.createElement('script');
+			simulateExtKeySequence.src = "../src/jquery.simulate.key-sequence.js";
+			iFrameDoc.body.appendChild(simulateExtKeySequence);
+			
+			setTimeout(function() {
+				notStrictEqual(window.frames[0].jQuery.simulate.prototype.quirks.delayedSpacesInNonInputGlitchToEnd, undefined, "Quirk detection was executed");
+				start();
+			}, 2000);
+		}, 1000);
+	});
+	
+	// See issue #9 (https://github.com/j-ulrich/jquery-simulate-ext/issues/9)
+	asyncTest("disable quirk detection", 1, function() {
+		setTimeout(function() {
+			var iFrameDoc = window.frames[0].document;
+	
+			var jquery = iFrameDoc.createElement('script');
+			jquery.src = "../libs/jquery-1.10.2.js";
+			iFrameDoc.body.appendChild(jquery);
+	
+			var bililiteRange = iFrameDoc.createElement('script');
+			bililiteRange.src = "../libs/bililiteRange.js";
+			iFrameDoc.body.appendChild(bililiteRange);
+			
+			var jquerySimulate = iFrameDoc.createElement('script');
+			jquerySimulate.src = "../libs/jquery.simulate.js";
+			iFrameDoc.body.appendChild(jquerySimulate);
+			
+			setTimeout(function() {
+				window.frames[0].jQuery.simulate.ext_disableQuirkDetection = true;
+	
+				var simulateExt = iFrameDoc.createElement('script');
+				simulateExt.src = "../src/jquery.simulate.ext.js";
+				iFrameDoc.body.appendChild(simulateExt);
+	
+				var simulateExtKeySequence = iFrameDoc.createElement('script');
+				simulateExtKeySequence.src = "../src/jquery.simulate.key-sequence.js";
+				iFrameDoc.body.appendChild(simulateExtKeySequence);
+				
+				setTimeout(function() {
+					strictEqual(window.frames[0].jQuery.simulate.prototype.quirks.delayedSpacesInNonInputGlitchToEnd, undefined, "Quirk detection was disabled");
+					start();
+				}, 2000);
+			}, 2000);
+		}, 1000);
+	});
 
 });